% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Angular-functions.R
\name{filter.by.2nn}
\alias{filter.by.2nn}
\title{Identify points whose two nearest neighbours are in line with or perpendicular to each other}
\usage{
filter.by.2nn(pts)
}
\arguments{
\item{pts}{An n-by-2 matrix containing the coordinates of the points to be compared.}
}
\value{
A Boolean array of length n that can be used to filter points or angles. TRUE indicates that the angle between the two nearest neighbours is closer to perpendicular than not; FALSE indicates that the absolute difference, modulo 2pi, is closer to 45 degrees than to 0 or 90.
}
\description{
For each point in the set, uses \code{k.nearest.angles} to obtain the 2 nearest neighbours, then identifies any for which the angle between them is more than pi/8 from perpendicular or parallel.
}
\examples{
k.2 <- k.nearest.angles(p, 2)
}

