% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Angular-functions.R
\name{k.nearest.angles}
\alias{k.nearest.angles}
\title{Get angle from each point to its k nearest neighbours}
\usage{
k.nearest.angles(pts, k)
}
\arguments{
\item{pts}{A two-column matrix containing the coordinates of the points to be compared.}

\item{k}{Number of neighbours to be included for each point.}
}
\value{
A (k+2)-column matrix; the first two columns contain each point's x and y coordinates, the remainder containing the angles found
}
\description{
For each point in the set, uses \code{knnx.index} to obtain the k nearest neighbours, then applies \code{atan2} to get the angle between them.
}
\examples{
k.2 <- k.nearest.angles(p, 2)
}

