% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Angular-functions.R
\name{pointwise.angles.atan2}
\alias{pointwise.angles.atan2}
\title{Extract angles between all points in a certain radius}
\usage{
pointwise.angles.atan2(pts, radius)
}
\arguments{
\item{pts}{A two-column matrix containing the coordinates of the points to be compared.}

\item{radius}{The radius (in the coordinate system of the plan) defining the area around each point within which the angles will be calculated.}
}
\value{
A (n x n+2) matrix containing the x, y coordinates in its first two columns, and the angles measured to every other point in the remaining columns. If no angle was measured, NA is returned in the appropriate cell.
}
\description{
Takes a set of x, y coordinates and, using \code{atan2}, returns a matrix of the angles from each point defined to all points within a specified radoius. (If no radius is specified, the angles from each point to every other point are returned)
}
\examples{
r0.2 <- pointwise.angles.atan2(p, radius = 5)
}

