% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Angular-functions.R
\name{show.directions}
\alias{show.directions}
\title{Display angles between nearest neighbours}
\usage{
show.directions(angles, l = 0.01)
}
\arguments{
\item{angles}{A matrix containing x and y coordinates in its first column, and angles to be plotted around each point in the remaining columns.}

\item{l}{Length of lines to be plotted}
}
\description{
Plot points & display the angle to each point's nearest neighbours
}
\examples{
show.directions(k.2)
}

