% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Angular-functions.R
\name{show.selected}
\alias{show.selected}
\title{Highlight a given point and all points within a certain radius}
\usage{
show.selected(angles, n = 1)
}
\arguments{
\item{angles}{A matrix containing x and y coordinates in its first column, and angles to be plotted around each point in the remaining columns.}

\item{n}{The index (row number) of the point to be plotted.}
}
\description{
All points in the set are plotted; the specified point is highlighted in red, and all points to which angles were measured are highlighted in blue. Allows the user to assess the number of points measured by \code{pointwise.angles.atan2} for diffrerent radii.
}
\examples{
show.selected(r0.2, n = 0)
}

